implementation apiService finished (23.12.23)

+++ app/src/main/java/com/kashapovrush/cofeeshopstest/data/model/Token.kt
@@ -4,5 +4,5 @@ import com.google.gson.annotations.SerializedName
 
 data class Token(
     @SerializedName("token") val token: String,
-    @SerializedName("tokenLifetime") val lifeTime: Long
+    @SerializedName("tokenLifetime") val lifeTime: Int
 )
